+++ e2e/test-env/src/client/RicosTestApp/utils.js
@@ -1,0 +1,37 @@
+export const createPresets = plugins => {
+  const partialPreset = [
+    plugins.image,
+    plugins.gallery,
+    plugins.video,
+    plugins.html,
+    plugins.divider,
+    plugins.spacing,
+    plugins.link,
+    plugins.hashtag,
+    plugins.mentions,
+    plugins.codeBlock,
+    plugins.soundCloud,
+    plugins.giphy,
+    plugins.headers,
+    plugins.map,
+    plugins.fileUpload,
+    plugins.linkButton,
+    plugins.textColor,
+    plugins.emoji,
+    plugins.highlight,
+    plugins.undoRedo,
+  ].filter(val => !!val); //non-mutual plugins such as undoRedo
+  return {
+    all: Object.values(plugins),
+    partialPreset,
+    embedsPreset: [plugins.link, plugins.linkPreview, plugins.verticalEmbed],
+    textPlugins: [
+      plugins.linkPreview,
+      plugins.verticalEmbed,
+      plugins.indent,
+      plugins.actionButton,
+      ...partialPreset,
+    ],
+    ...plugins,
+  };
+};