+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -39,6 +39,7 @@ import draftStyles from '../../statics/styles/draft.rtlignore.scss';
 import 'wix-rich-content-common/dist/statics/styles/draftDefault.rtlignore.scss';
 import { convertFromHTML as draftConvertFromHtml } from 'draft-convert';
 import { pastedContentConfig, clearUnnecessaryInlineStyles } from './utils/pastedContentUtil';
+import { deprecateHelpers } from 'wix-rich-content-common/dist/lib/deprecateHelpers.cjs.js';
 
 class RichContentEditor extends Component {
   static getDerivedStateFromError(error) {
@@ -133,7 +134,7 @@ class RichContentEditor extends Component {
       iframeSandboxDomain,
     } = this.props;
 
-    this.fixFileHandlersName(helpers);
+    this.fixHelpers(helpers);
 
     this.contextualData = {
       theme: theme || {},
@@ -262,16 +263,17 @@ class RichContentEditor extends Component {
     if (this.props.textToolbarType !== nextProps.textToolbarType) {
       this.setState({ textToolbarType: nextProps.textToolbarType });
     }
-    this.fixFileHandlersName(nextProps.helpers);
+    this.fixHelpers(nextProps.helpers);
   }
 
-  fixFileHandlersName(helpers) {
+  fixHelpers(helpers) {
     if (helpers?.onFilesChange) {
       // console.warn('helpers.onFilesChange is deprecated. Use helpers.handleFileUpload');
       helpers.handleFileUpload = helpers.onFilesChange;
       // eslint-disable-next-line fp/no-delete
       delete helpers.onFilesChange;
     }
+    deprecateHelpers(helpers, this.props.config);
   }
 
   // TODO: get rid of this ASAP!