+++ packages/plugin-gallery/web/src/gallery-viewer.jsx
@@ -11,6 +11,7 @@ import ExpandIcon from './icons/expand.svg';
 import { GALLERY_TYPE } from './types';
 
 const { ProGallery, GALLERY_CONSTS } = require('pro-gallery');
+import 'pro-gallery/dist/statics/main.css';
 
 class GalleryViewer extends React.Component {
   constructor(props) {
@@ -24,7 +25,7 @@ class GalleryViewer extends React.Component {
   }
 
   componentDidMount() {
-    if (this.props.helpers.onExpand) {
+    if (this.props.settings.onExpand) {
       const styleParams = this.state.styleParams;
       this.setState({
         styleParams: { ...styleParams, allowHover: true },
@@ -39,7 +40,7 @@ class GalleryViewer extends React.Component {
     if (!scrollingElement) {
       // eslint-disable-next-line no-console
       console.error(
-        `Please fix the gallery config of Rich Content Editor. 
+        `Please fix the gallery config of Rich Content Editor.
         A scrollingElement needs to be provided. Without it the gallery will not work correctly`
       );
       scrollingElement = document.body;
@@ -143,8 +144,11 @@ class GalleryViewer extends React.Component {
   };
 
   handleExpand = data => {
-    const { onExpand, onAction } = this.props.helpers;
-    onAction?.('expand_gallery', GALLERY_TYPE);
+    const {
+      settings: { onExpand },
+      helpers = {},
+    } = this.props;
+    helpers.onAction?.('expand_gallery', GALLERY_TYPE);
     onExpand?.(this.props.entityIndex, data.idx);
   };