+++ packages/viewer/web/src/utils/convertContentState.jsx
@@ -1,13 +1,17 @@
 import React from 'react';
 import { renderToStaticMarkup } from 'react-dom/server';
-import { BLOCK_TYPES, depthClassName } from 'wix-rich-content-common';
+import {
+  BLOCK_TYPES,
+  depthClassName,
+  getTextDirection,
+  getDirectionFromAlignmentAndTextDirection,
+} from 'wix-rich-content-common';
 import redraft from 'wix-redraft';
 import classNames from 'classnames';
 import { endsWith } from 'lodash';
 import List from '../List';
 import getPluginViewers from '../getPluginViewers';
 import { kebabToCamelObjectKeys } from './textUtils';
-import { getTextDirection } from './textDirection';
 import { staticInlineStyleMapper } from '../staticInlineStyleMapper';
 import { combineMappers } from './combineMappers';
 import { getInteractionWrapper, DefaultInteractionWrapper } from './getInteractionWrapper';
@@ -24,7 +28,11 @@ const getBlockDepth = (contentState, key) =>
   contentState.blocks.find(block => block.key === key).depth || 0;
 
 const getBlockStyleClasses = (data, mergedStyles, textDirection, classes) => {
-  const rtl = textDirection === 'rtl' || data.textDirection === 'rtl';
+  const rtl =
+    getDirectionFromAlignmentAndTextDirection(
+      data.textAlignment,
+      textDirection || data.textDirection
+    ) === 'rtl';
   const defaultTextAlignment = rtl ? 'right' : 'left';
   const alignmentClass = data.textAlignment || defaultTextAlignment;
   return classNames(